Add files via upload
This zip file contains all the parks for Roller Coaster Tycoon 1, saved a few minutes before winning. Simply open them up in numerical order, let the game win, and move on to the next one. There is also a MegaPark with flat land and millions of dollars, not created with any cheats. No Drexler or other trainers required.